+++ popOsSetup.sh
@@ -4,34 +4,54 @@
 tempDir = $(mktemp -d)
 trap 'rm -rf "$tempDir"' EXIT
 
-#Update system
-sudo apt update 
-sudo apt upgrade -y
-sudo apt-get update
-sudo apt-get upgrade -y
-
-#some system installation stuff
+if command -v apt >/dev/null 2>&1; then 
+    #Update system
+    sudo apt update 
+    sudo apt upgrade -y
+    sudo apt-get update
+    sudo apt-get upgrade -y
 
 #install ranger a cli file manager
-sudo apt install ranger -y
+    sudo apt install ranger -y
 
 #unistall Firefox Browswer
-sudo apt-get purge firefox
-sudo rm -Rf /etc/firefox/
+    sudo apt-get purge firefox
+    sudo rm -Rf /etc/firefox/
+    sudo rm -rf ~/.mozilla/firefox
+
+    #removal of snap
+    curl -sSfL https://raw.githubusercontent.com/edickens09/New-system-setup/main/snapRemove.sh -o "$tempDir/snapRemove.sh"
+
+    if [ -s "$tempDir/snapRemove.sh"]; then
+        source "$tempDir/snapRemove.sh"
+    else
+        echo "Download failed"
+	exit 1
+    fi
+
+fi
+
+if command -v dnf >/dev/null 2>&1; then 
+   #update system 
+    sudo dnf upgrade -y
+
+    #uninstall Firefox Browser
+    sudo dnf remove firefox
+    rm -rf ~/.mozilla/firefox
+
+fi
+
 
 #install zoxide which is used to replace cd
-curl -sSfL https://raw.githubusercontent.com/ajeetdsouza/zoxide/main/install.sh | sh
-echo 'eval "$(zoxide init bash)"' >> ~/.bashrc
-source ~/.bashrc
+curl -sSfL https://raw.githubusercontent.com/edickens09/New-system-setup/main/zoxide.sh -o "$tempDir/zoxide.sh"
 
-#removal of snap
-curl -sSfL https://raw.githubusercontent.com/edickens09/New-system-setup/main/snapRemove.sh -o "$tempDir/snapRemove.sh"
+if [ -s "$tempDir/zoxide.sh" ]; then
+    source "tempDir/zoxide.sh"
 
-if [ -s "$tempDir/snapRemove.sh"]; then
-	source "$tempDir/snapRemove.sh"
 else
-	echo "Download failed"
-	exit 1
+    echo "Download failed"
+    exit 1
+
 fi
 
 #Install Go
@@ -87,16 +107,12 @@ fi
 mkdir -p ~/Workspace/Github/edickens09
 mkdir -p ~/Workspace/Gitea/eric
 
-#setup Neovim config
-#should I do this or should I wait until git is setup and then clone git repo?
-#mkdir -p ~/.config/nvim/lua/{config, plugins}
-#touch ~/.config/nvim/init.lua
-#touch ~/.config/nvim/lua/config/lazy.lua
-
 #sourceing the new .bashrc to have new command take effect
 source ~/.bashrc
 
-#Cleanup
-sudo apt autoremove -y
+if command -v apt >/dev/null 2>&1; then
+    #Cleanup
+    sudo apt autoremove -y
+fi
 
 exec "$SHELL"
